added states handing to single query

+++ aro-service/aro-engine/src/main/java/com/altvil/aro/service/processing/UserProcessingLayerService.java
@@ -11,7 +11,7 @@ public interface UserProcessingLayerService {
     Collection<ServiceLayer> getUserServiceLayers(int userId) ;
     ServiceLayer getUserServiceLayers(int userId, int id) ;
 
-    ServiceLayer addUserServiceLayer(Long userId, String layerName, String layerDescription);
+    ServiceLayer addUserServiceLayer(int userId, String layerName, String layerDescription);
 
     void loadUserServiceLayerEntitiesCSV(int id, Writer responseWriter);
